src/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
clipboard==0.0.4
idna==3.4
jeepney==0.8.0
loguru==0.6.0
notify-py==0.3.42
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pyperclip==1.8.2
pytesseract==0.3.10
requests==2.28.2
urllib3==1.26.14
